added "script" identification
It's based on Unicode blocks used by a text.  If I compute the set of
blocks used by each language in the catalog, I can ignore languages that
are missing blocks used by the target text.  (I'll also need to compute
a list of special symbol blocks used by *none* of the languages, so that
I can ignore them in the pruning)

...

Actually, these blocks are probably only useful in some sort of
frequency list, so that only the most prominant blocks can prune out
languages, and not scattered other symbols (greek mathematical symbols,
for instance)