new features / metafeatures established
features are a single hop in a tree of (feature -> feature) with no
cycles

metafeatures are an automatically derived (hopefully?) hop made of any
number of feature hops, as well as any other arbitrary hops I'd like to
make (such as ones in the opposite direction of existing feature hops,
where such a thing is possible)

also tokens have been lumped in with regular features - there is
opportunity here to split them into a chain of several features, such
as:

    CharSeq -> OrderedTokenList -> TokenList -> TokenSet

This would be useful for specifying exactly what information a
particular feature needs or can be converted back to.  (A trigram
frequency-list can be converted to a TokenList without losing
information, but *not* all the way back to an OrderedTokenList.
It's useful to know this, because a UBlock set only needs a TokenSet,
which can be derived from that TokenList)