added cosineX function
This is like cosineM, but allows for many languages to be cosine'd
simultaneously (meaning, each call to the fetchFrequency function passed
in returns the frequency of the current trigram for every langyage being
considered).

This should cut down on the number of database queries required while
still only requesting trigrams that are actually needed.